+++ src/pretalx/person/forms.py
@@ -142,7 +142,7 @@ class SpeakerProfileForm(
             kwargs["instance"] = self.user.event_profile(self.event)
         else:
             kwargs["instance"] = SpeakerProfile()
-        super().__init__(*args, **kwargs, event=self.event)
+        super().__init__(*args, **kwargs, event=self.event, limit_to_rooms=True)
         read_only = kwargs.get("read_only", False)
         initial = kwargs.get("initial", dict())
         initial["name"] = name